pyproject.toml
[project]
name = "TeCoD"
version = "0.2.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.1.1",
    "dill>=0.4.0",
    "openai>=1.0.0",
    "hydra-core>=1.3.2",
    "ipywidgets>=8.1.7",
    "numpy>=2.3.3",
    "outlines<1.2.0",
    "pandas>=2.3.3",
    "pydantic>=2.8.0",
    "pymilvus[milvus-lite]>=2.6.2",
    "sentence-transformers>=5.1.1",
    "sqlglot>=27.21.0",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "typer>=0.19.2",
]

[project.scripts]
tecod = "main:app"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
only-include = ["src", "main.py"]
exclude = [
    "**/.git",
    "**/.gitignore",
    "**/.gitmodules",
    "**/__pycache__",
    "**/*.pyc",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
lint = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "black>=24.0.0",
    "pre-commit>=3.7.0",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "src/__init__.py",
    "src/*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["src/pdec"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "W",      # pycodestyle warnings
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]
